Changed to allow variable param values per compartment
VERTEX has the capability to allow parameters C, R_M, R_A and E_leak to
be specified on a per-compartment basis, but the checks on the user
inputs used to prevent this. This has been modified in the
checkNeuronStruct.m file to allow these parameters to be specified
either as scalars or vectors.