Adds ability to "stretch" neurons
Can now specify to stretch neurons up to the top of the model (z-axis)
to simulate different lengths of pyramidal cells depending on their
soma position. This currently only affects calculation of the LFP, it
does not change the neuron’s passive dynamics at all (these are still
calculated from the dimensions given in the NeuronParams structure).
Compartment positions are now stored in
params.TissueParams.compartmentPositions.